remove invoices duplicates
