+++ app/src/server/scripts/removeDuplicateInvoices.ts
@@ -1,8 +1,4 @@
-import dayjs from "dayjs";
-import { kysely, prisma } from "../db";
-
-const startOfFeb2024 = dayjs("2024-02-01").startOf("month").toDate();
-const endOfFeb2024 = dayjs("2024-02-01").endOf("month").toDate();
+import { prisma } from "../db";
 
 const invoices = await prisma.invoice.findMany({
   where: {
@@ -36,23 +32,4 @@ for (const invoice of invoices) {
     });
     deletedInvoiceIds.add(duplicateInvoice.id); // Add the deleted invoice ID to the set
   }
-}
-
-const invoices2 = await prisma.invoice.findMany({
-  where: {
-    billingPeriod: "Feb 2024",
-  },
-});
-
-for (const invoice of invoices2) {
-  await kysely
-    .updateTable("UsageLog")
-    .set({
-      invoiceId: invoice.id,
-    })
-    .where("projectId", "=", invoice.projectId)
-    .where("createdAt", ">=", startOfFeb2024)
-    .where("createdAt", "<=", endOfFeb2024)
-    .where("invoiceId", "is", null)
-    .execute();
 }
